23.11.2023

+++ openserver-sites/shop.loc/app/views/parts/products_loop.php
@@ -16,7 +16,7 @@
                             </div>
 							<div class="product-links">
 								<a class="add-to-cart" href="cart/add?id=<?= $product['id'] ?>" data-id="<?= $product['id'] ?>"><?= get_cart_icon($product['id']) ?></i></a>
-                                <?php if(in_array($product['id'], \wfm\App::$app->getProperty('wishlist'))): ?>
+                                <?php if((\wfm\App::$app->getProperty('wishlist')) && in_array($product['id'], \wfm\App::$app->getProperty('wishlist'))): ?>
                                     <a class="delete-from-wishlist" href="wishlist/delete?id=<?= $product['id'] ?>" data-id="<?= $product['id'] ?>"><i class="fas fa-hand-holding-heart"></i></a>
                                 <?php else: ?>
                                     <a class="add-to-wishlist" href="wishlist/add?id=<?= $product['id'] ?>" data-id="<?= $product['id'] ?>"><i class="far fa-heart"></i></a>
